backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.0
deep-translator==1.11.0
elevenlabs==1.58.1
fastapi==0.115.12
Flask==2.2.2
frozenlist==1.6.0
gevent==25.4.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.3
openai==0.27.0
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
python-dotenv==0.21.0
requests==2.28.2
setuptools==80.4.0
simpleaudio==1.0.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.34.2
websocket==0.2.1
websockets==15.0.1
Werkzeug==2.2.2
yarl==1.20.0
zope.event==5.0
zope.interface==7.2
